+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.1] - 2023-02-04
 ### Added
 - Added buttons to add and remove pets, users to and from groups
+- Added edit group page
 ### Changed
 - Fixed My Pets and My Groups to accurately reflect only the pets and groups that the user owns/belongs to
 - Fixed Add Pets to Group to only include user's pets
+- Members of each item type (pet users, group pets, etc.) now link back to those items
 
 ## [0.6.0] - 2023-01-19
 ### Added